Work around clojure.instant bootclasspath problem
As of Java 9, certain build tools during boot will use the bootclasspath
(this includes both Lein and Boot afaik). This is a problem for
clojure.instant as it relies on java.sql.Timestamp which isn't available
on the bootclasspath.

With Fipp, this means that libraries and plugins requiring fipp.ednize
during the build system's operation will fail unless LEIN_USE_BOOTCLASSPATH
is set to no (or similar workaround for Boot).

This (admittedly somewhat evil) commit circumvents that by checking to
see if clojure.instant has already been loaded and only extending the
edn protocol for date/time related objects if it has been loaded.