Special handling for java.sql.Date
java.sql.Date is a subclass of java.util.Date but throws a runtime
exception `UnsupportedOperationException` when called, because a SQL
date inherently has no time information associated with it. It is just
a date.

Therefore a special case is added to the protocol to support pretty
printing java.sql.Date objects.